Reduce Travis build matrix
There is no need to test each compiler/architecture combination using
both build systems. This commit uses CMake as the primary build system
for Travis (since it is significantly faster) and includes one autotools
build for a single compiler/architecture to verify that the autotools
system can continue to be used.