Move CoordinateSequence::add variants to CoordinateArraySequence
Making this change removes the assumption that any CoordinateSequence is
resizable, which opens the door to a CoordinateSequence that is blocked
by a raw block of memory (such as a PostGIS POINTARRAY).

The default CoordinateArraySequence constructor now creates a
CoordinateSequence of unknown dimension (dimension determined when the
first Coordinate is assigned) rather than of dimension 3. This is for
consistency with the behavior of the CoordinateArraySequenceFactory.