Remove Puntal, Lineal, and Polygonal classes
These classes are used in only a handful of locations to check the
dimension of a Geometry. However, their use in the inheritance hierarchy
increases the class size of MultiPoint, MultiLineString, and
MultiPolygon by 8 bytes, and possibly imposes a runtime cost. This commit
provides alternative methods of checking Geometry dimension.