Optimize IndexedNesterRingTester
- Use IndexedPointInAreaLocator for ring containment tests
- Use envelope covers test instead of intersection test
- Reuse results vector across iterations
- Presize ring vector

The first change can bring a significant performance improvement
polygons that have many or complex rings. An example is a parcel or land
cover dataset that models the street network as a single polygon with
many holes. It's possible that the indexed locator could perform worse
in some cases but I did not find any in my testing.

The last three changes should have no adverse effect in any situation.