Fix -Wmismatched-tags warning related to WK[T/B][Reader/Writer]
CLang 9 -Wextra warns about

```
./geos_c.h:1226:9: warning: struct 'WKTReader' was previously declared as a class; this is valid, but may result in linker errors under the Microsoft C++ ABI [-Wmismatched-tags]
typedef struct GEOSWKTReader_t GEOSWKTReader;
        ^
../../include/geos/io/WKTReader.h:58:16: note: previous use is here
class GEOS_DLL WKTReader {
               ^
./geos_c.h:1226:9: note: did you mean class here?
typedef struct GEOSWKTReader_t GEOSWKTReader;
        ^~~~~~
        class
```
Similarly for WKTWriter, WKBReader and WKBWriter

This is due to capi/geos_c.cpp having a
```
```
Hence when geos_c.h is included from geos_c.cpp, it is expanded as
typedef struct geos::io::WKTReader GEOSWKTReader;
which conflicts with the class GEOS_DLL WKTReader declaration.

The easy fix is to change the declaration of WKTReader to be a struct.
As the visibility is set to public immediately, this is mostly a no-op
change, and fixes the warning.

Same for the other 3 classes/structs