Remove default arguments for Coordinate constructor
The previously defined constructor allow a Coordinate to be constructed
from only a single value, i.e. Coordinate(3) would produce (3, 0, NaN).
It's not clear what the use for this is, and it allowed further
complications like Envelope(3) compiling by implicitly constructing the
coordinate above.

This commit adds a no-argument constructor to preserve the behavior of
Coordinate() producing (0, 0, NaN).