Avoid out-of-memory for malformed WKB.
By simply creating a WKB input with a vertex count/ring count/geometry count that is very very large, a user can get GEOS to allocate a very large vector in advance of actually parsing the WKB to fill in that vector.

Instead, we do a quick bounds check, comparing the number of objects requested and the minimum size those objects could be with the amount of space left in the input stream, and raising an exception if the minimum possible request size exceeds the amount of input data.