Add methods to release objects owned by some Geometry types
This can allow for more efficient implementation implementation of some
operations such as GeometryCombiner. Objects that release their
components are left in an undefined state. This approach is taken
because it is analogous to the state of moved-from objects, and because
reverting the objects to a defined state can be expensive. For example,
an empty Polygon must own an empty LinearRing, which must own an empty
CoordinateSequence.