Invert conditional in unionSafe()
This prevents the compiler from mistakenly thinking that the
dereferenced geometry is .  The compiler *should* have been able to
figure this out and suppress the warning but it doesn't.

This change shouldn't affect the functionality of the code.  It only
gives a better hint to the compiler about the validity of the pointer
to Geometry.