Don't initialize CoordinateSequence in constructor
This ought to improve performance in the common case where
initialization is unnecessary.

However, the `getDimension` method, in the case where the dimension is
unknown (0) looks at the z dimension of the first element in the array
to determine if there is z.  So we need to initialize at least the
first element.