CoordinateFilter: Support multiple CoordinateSequence dimensions
The interface of CoordinateFilter is changed so that a filter should
support operations on CoordinateXY* rather than Coordinate*. By default,
higher-dimension coordinates will default to base-class implementations
(e.g., a CoordinateXYZM will be processed as a CoordinateXY if no
implementation is available for Coordiante or CoordinateXYZM.)

To avoid the case where the filter implementation does not need to
consider Z or M but must preserve them (e.g., repeated point removal),
the CoordinateInspector and CoordinateMutator base classes are provided.
These allow a filter class to inherit from the base using CRTP and
provide a template for the filter method rather than an implementation
for each coordinate type.