Improve OffsetCurve behaviour and add Joined mode.
Port of https://github.com/locationtech/jts/pull/956

The output contains as much linework as possible that can be determined to be in the offset curve (based on the buffer boundary of the input). This will be a single line in simple cases (which matches the previous implementation, and the old GEOS code). For more complex cases containing "narrow corridors" or self-intersections the output will be a MultiLineString containing the various disjoint curves which lie at the given offset distance from the input line. The output is structured to make it possible to reduce if necessary; in particular, the line elements are ordered along the raw offset curve as much as possible, with the main or outer curve appearing first.