Port Coverage Simplification https://github.com/locationtech/jts/pull/911
Adds the CoverageSimplifier class to the coverage package. This supports simplification of the edges of polygonal coverages while preserving the coverage topology:

    The simplified coverage has the same number and types of polygonal geometries as the input
    Coverage node points (inner vertices shared by three or more polygons, or boundary vertices shared by two or more) are not changed
    if the input is a valid coverage, then so is the result

The simplification uses an approach equivalent to Visvalingam-Whyatt simplification. This is based on removing vertices which are the apex of low-area "corners". The amount of simplification is determined by a tolerance value which is the square root of the maximum corner area that can be removed.