Overlay with mixed dimension (#923)
For overlay operations (union/difference/intersection/symdifference) support for GeometryCollection has been missing, at least for those cases of GeometryCollection that contain a mixture of types (point/line/polygon).

This fix tests inputs to HeuristicOverlay (formerly a wide-ranging set of shims to work around issues with the old overlay code) and for those that are mixed-type collections, processes the component dimensions separately, with appropriate logic for the overlay operation being run. Fixes GH-797.