Support floatint-point id
Use "dump()" instead of get<int>() or get<float>() to make code clear.
And notice that both 123.0 and 123.000 will be parsed as "123.0" because
nlohmann/json will cut the suffix 0 by default. It is hard to support
float well. But I think use floating-point as id is not a good practice.
So I think then current implementation is acceptable.

Signed-off-by: Kunlin Yu <yukunlin@syriusrobotics.com>